Move JWK decoding to a separate class
